+++ c/formats/ted5.h
@@ -73,7 +73,7 @@ void ted5_close(ted5_t *ted5);
 
 /* get size (in bytes) of a given map's plane, for memory allocation */
 /* returns 0 on error */
-uint32_t ted5_get_plane_size(ted5_t *ted5, int map, int plane);
+size_t ted5_get_plane_size(ted5_t *ted5, int map, int plane);
 
 /* read map plane into user-allocated buffer */
 /* allocated buffer size should be determined with ted5_get_plane_size() */
